Changed return value logic for GetUserSongs
We understood the assignment wrong and returned IResult in out handler methods. The handlers shouldn't have anything to do with the web or the api, only interract with the db. So changed the lgoic so the hanlder method returns the object we want and in the handler we return the IResult to the request.

Need to change all methods to this strucure