Moved SpotifyHandler code here.
Moved all code from SpotifyHandler  to its own class SpotifyHelper to be consistent with the rest of the programs structure.

This class consists of the interface and class. The constructors that create the HttpClient.

Methods to get token and access token from Spotify's API.

And now a search method that takes in a search query and optional parameters offset and country code. Returns 10 best song results matching query.
Results consist of song ID, song name, list of artists and their ID's.