Fixed return type  and logic of AddSong
Changed return type to void as we don't need to return anything only create a song with this method.
Input parameters have changed from the route to a JSON file. So removed those and changed to a DTO. Thus we could remove taking out the artist and genre ids as that's unecessary now